fixed a bug where the value of x was typed to be an int and was overflowing to be a negative number in the calculation of sigma
git-svn-id: svn://orca.physics.unc.edu/Orca@2901 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Widgets/Plotting/ORGate1D.m
@@ -231,13 +231,13 @@ const float kGateAlpha2 = .1;
         double sumXY = 0.0;
         double sumX2Y = 0.0;
 		double sumValX = 0;
-		int maxX = 0;
+		float maxX = 0;
 		float minY = 3.402e+38;
 		float maxY = -3.402e+38;
-        float	x;
+        float x;
 		float y;
 		int xStart = [self gateMinChannel];
-		int xEnd = [self gateMaxChannel];
+		int xEnd   = [self gateMaxChannel];
 		
 		int totalNum = xEnd - xStart+1;
